nix: workaround bug with dependencies of packagesExtra
There is a bug in mach-nix where using `packagesExtra` with
`mkPython`: dependencies of extra packages are not merged with
dependencies computed from requirements. This means two different
versions of a package can be installed and lead to a collision. In our
case, ensure we use the "right" version of pulumi in packages in
packagesExtra.

More details in https://github.com/DavHau/mach-nix/issues/327.